Day 25. A sligtly optimized version
We know definitely that the one of the edges to cut will be in the spanning tree.
And with a high probability one of the remaining edges will be not in the spanning tree.
That allows us to iterate over all possible pair of the two first edges faster. The final
part is finding bridges in the graph with two cut edges.